chore: update code structure according to review comments
