+++ src/electron/windows/digicert-usb-config/eToken-macos.cfg
