+++ src/electron/windows/digicert-usb-config/eToken-windows.cfg
